design leaderboard

+++ app/src/main/java/com/example/annotatex_mobile/LeaderboardItem.java
@@ -2,11 +2,11 @@ package com.example.annotatex_mobile;
 
 public class LeaderboardItem {
     private String userName;
-    private int booksRead;
+    private long booksRead;
     private String profileImageUrl;
     private boolean isCurrentUser;
 
-    public LeaderboardItem(String userName, int booksRead, String profileImageUrl, boolean isCurrentUser) {
+    public LeaderboardItem(String userName, long booksRead, String profileImageUrl, boolean isCurrentUser) {
         this.userName = userName;
         this.booksRead = booksRead;
         this.profileImageUrl = profileImageUrl;
@@ -14,7 +14,7 @@ public class LeaderboardItem {
     }
 
     public String getUserName() { return userName; }
-    public int getBooksRead() { return booksRead; }
+    public long getBooksRead() { return booksRead; }
     public String getProfileImageUrl() { return profileImageUrl; }
     public boolean isCurrentUser() { return isCurrentUser; }
 }
